Fixed the routing issue
