+++ package.json
@@ -1,63 +1,4 @@
 {
-<<<<<<< HEAD
-  "name": "rondwell-project-frontend",
-  "private": true,
-  "version": "0.0.2",
-  "type": "module",
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "vite preview",
-    "prepare": "svelte-kit sync || echo ''",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "format": "prettier --write .",
-    "lint": "prettier --check . && eslint ."
-  },
-  "devDependencies": {
-    "@eslint/compat": "^1.2.5",
-    "@eslint/js": "^9.18.0",
-    "@sveltejs/adapter-auto": "^6.0.0",
-    "@sveltejs/adapter-static": "^3.0.8",
-    "@sveltejs/kit": "^2.16.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.0",
-    "@tailwindcss/forms": "^0.5.9",
-    "@tailwindcss/postcss": "^4.1.7",
-    "@tailwindcss/typography": "^0.5.15",
-    "@tailwindcss/vite": "^4.0.0",
-    "autoprefixer": "^10.4.21",
-    "eslint": "^9.18.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-svelte": "^3.0.0",
-    "globals": "^16.0.0",
-    "postcss": "^8.5.3",
-    "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.6",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "svelte": "^5.0.0",
-    "svelte-check": "^4.0.0",
-    "svelte-motion": "^0.12.2",
-    "tailwind-variants": "^0.3.0",
-    "tailwindcss": "^4.1.7",
-    "typescript": "^5.0.0",
-    "typescript-eslint": "^8.20.0"
-  },
-  "dependencies": {
-    "@carbon/charts-svelte": "^1.22.18",
-    "@iconify/svelte": "^5.2.1",
-    "@sveltejs/adapter-netlify": "^5.0.2",
-    "bits-ui": "^0.21.2",
-    "carbon-components": "^10.59.2",
-    "d3": "^7.9.0",
-    "gsap": "^3.13.0",
-    "lucide-svelte": "^0.454.0",
-    "motion-svelte": "^0.0.1",
-    "svelte-intersection-observer": "^1.0.0",
-    "tailwindcss-animate": "^1.0.7",
-    "vite": "^6.2.6",
-    "vite-plugin-pwa": "^1.0.0"
-  }
-=======
 	"name": "rondwell-project-frontend",
 	"private": true,
 	"version": "0.0.2",
@@ -89,9 +30,9 @@
 		"eslint-plugin-svelte": "^3.0.0",
 		"globals": "^16.0.0",
 		"postcss": "^8.5.3",
-		"prettier": "^3.4.2",
-		"prettier-plugin-svelte": "^3.3.3",
-		"prettier-plugin-tailwindcss": "^0.6.11",
+		"prettier": "^3.2.5",
+		"prettier-plugin-svelte": "^3.2.6",
+		"prettier-plugin-tailwindcss": "^0.5.11",
 		"svelte": "^5.0.0",
 		"svelte-check": "^4.0.0",
 		"svelte-motion": "^0.12.2",
@@ -115,5 +56,4 @@
 		"vite": "^6.2.6",
 		"vite-plugin-pwa": "^1.0.0"
 	}
->>>>>>> origin/dev
 }
