+++ src/lib/stores/uiStore.ts
@@ -9,8 +9,3 @@ export interface SubMenuItem {
 export const showSubMenu = writable(false);
 export const subMenuItems = writable<SubMenuItem[]>([]);
 export const activeSubItem = writable('');
-
-// Settings subMenu
-export const showSettingsSubMenu = writable(false);
-export const settingsSubMenuItems = writable<SubMenuItem[]>([]);
-export const activeSettingsItem = writable('');
