--- src/routes/(app)/(main)/collection/[id]/+layout.server.ts
@@ -1,1 +1,0 @@
-export const prerender = false;
