+++ src/routes/(app)/(main)/collection/[id]/+layout.svelte
@@ -1,20 +1,12 @@
 <script lang="ts">
-	
+	import { browser } from '$app/environment';
 	import { page } from '$app/stores';
-	import {
-	showSubMenu,
-	subMenuItems,
-	activeSubItem,
-	showSettingsSubMenu,
-	settingsSubMenuItems,
-	activeSettingsItem
-
-} from '$lib/stores/uiStore';
-
+	import { activeSubItem, showSubMenu, subMenuItems } from '$lib/stores/uiStore.js';
+	import { onMount } from 'svelte';
 	import { get } from 'svelte/store';
-	
+	import Sidebar from '../../../components/Sidebar.svelte';
+	import SideMenu from '../../../components/SideMenu.svelte';
 
-	
 	const collectionIcon = {
 		eventIcon: `<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
 <path d="M13.9081 18.9584H6.09144C2.61644 18.9584 1.55811 17.9 1.55811 14.425V14.0417C1.55811 13.7 1.84144 13.4167 2.18311 13.4167C2.91644 13.4167 3.51644 12.8167 3.51644 12.0834C3.51644 11.35 2.91644 10.75 2.18311 10.75C1.84144 10.75 1.55811 10.4667 1.55811 10.125V9.74171C1.55811 6.26671 2.61644 5.20837 6.09144 5.20837H13.9081C17.3831 5.20837 18.4414 6.26671 18.4414 9.74171V10.525C18.4414 10.8667 18.1581 11.15 17.8164 11.15C17.0831 11.15 16.4831 11.75 16.4831 12.4834C16.4831 13.2167 17.0831 13.8084 17.8164 13.8084C18.1581 13.8084 18.4414 14.0917 18.4414 14.4334C18.4331 17.9 17.3748 18.9584 13.9081 18.9584ZM2.81644 14.5917C2.83311 17.2417 3.35811 17.7084 6.09977 17.7084H13.9164C16.5164 17.7084 17.1164 17.2834 17.1914 14.975C16.0748 14.6917 15.2414 13.6834 15.2414 12.475C15.2414 11.2667 16.0748 10.25 17.1998 9.96671V9.73337C17.1998 6.94171 16.7081 6.45004 13.9164 6.45004H6.09144C3.35811 6.45004 2.83311 6.92504 2.80811 9.56671C3.93311 9.85004 4.76644 10.8667 4.76644 12.075C4.76644 13.2834 3.93311 14.3084 2.81644 14.5917Z" fill="currentColor"/>
@@ -70,85 +62,36 @@
 </svg>`
 	};
 
+	let isMobile = false;
 
-	const settingIcon = {
-		displayIcon: `<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M13.9081 18.9584H6.09144C2.61644 18.9584 1.55811 17.9 1.55811 14.425V14.0417C1.55811 13.7 1.84144 13.4167 2.18311 13.4167C2.91644 13.4167 3.51644 12.8167 3.51644 12.0834C3.51644 11.35 2.91644 10.75 2.18311 10.75C1.84144 10.75 1.55811 10.4667 1.55811 10.125V9.74171C1.55811 6.26671 2.61644 5.20837 6.09144 5.20837H13.9081C17.3831 5.20837 18.4414 6.26671 18.4414 9.74171V10.525C18.4414 10.8667 18.1581 11.15 17.8164 11.15C17.0831 11.15 16.4831 11.75 16.4831 12.4834C16.4831 13.2167 17.0831 13.8084 17.8164 13.8084C18.1581 13.8084 18.4414 14.0917 18.4414 14.4334C18.4331 17.9 17.3748 18.9584 13.9081 18.9584ZM2.81644 14.5917C2.83311 17.2417 3.35811 17.7084 6.09977 17.7084H13.9164C16.5164 17.7084 17.1164 17.2834 17.1914 14.975C16.0748 14.6917 15.2414 13.6834 15.2414 12.475C15.2414 11.2667 16.0748 10.25 17.1998 9.96671V9.73337C17.1998 6.94171 16.7081 6.45004 13.9164 6.45004H6.09144C3.35811 6.45004 2.83311 6.92504 2.80811 9.56671C3.93311 9.85004 4.76644 10.8667 4.76644 12.075C4.76644 13.2834 3.93311 14.3084 2.81644 14.5917Z" fill="currentColor"/>
-<path d="M8.3335 8.54171C7.99183 8.54171 7.7085 8.25837 7.7085 7.91671V5.83337C7.7085 5.49171 7.99183 5.20837 8.3335 5.20837C8.67516 5.20837 8.9585 5.49171 8.9585 5.83337V7.91671C8.9585 8.25837 8.67516 8.54171 8.3335 8.54171Z" fill="currentColor"/>
-<path d="M8.3335 14.1C7.99183 14.1 7.7085 13.8166 7.7085 13.475V10.7C7.7085 10.3583 7.99183 10.075 8.3335 10.075C8.67516 10.075 8.9585 10.3583 8.9585 10.7V13.475C8.9585 13.8166 8.67516 14.1 8.3335 14.1Z" fill="currentColor"/>
-<path d="M8.3335 18.9583C7.99183 18.9583 7.7085 18.675 7.7085 18.3333V16.25C7.7085 15.9083 7.99183 15.625 8.3335 15.625C8.67516 15.625 8.9585 15.9083 8.9585 16.25V18.3333C8.9585 18.675 8.67516 18.9583 8.3335 18.9583Z" fill="currentColor"/>
-<path d="M13.6082 6.45828H6.03315C5.78315 6.45828 5.54982 6.30828 5.45815 6.07495C5.36649 5.84162 5.41649 5.56662 5.59149 5.39162L8.03315 2.94995C10.2832 0.699951 11.6582 0.699951 13.8998 2.94995L14.3998 3.44995C14.5165 3.56662 14.5832 3.72495 14.5832 3.89162C14.5832 4.05828 14.5165 4.21662 14.3998 4.33328C14.0665 4.66662 13.9748 5.15828 14.1665 5.58328C14.2582 5.77495 14.2415 5.99995 14.1248 6.18328C14.0165 6.34995 13.8165 6.45828 13.6082 6.45828ZM7.54149 5.20828H12.8165C12.7998 4.76662 12.9082 4.32495 13.1332 3.94162L13.0248 3.83328C11.2832 2.09162 10.6665 2.09162 8.92482 3.83328L7.54149 5.20828Z" fill="currentColor"/>
-</svg>`,
-		paymentIcon:`<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M12.4998 18.9583H7.49984C2.97484 18.9583 1.0415 17.025 1.0415 12.5V7.49999C1.0415 2.97499 2.97484 1.04166 7.49984 1.04166H12.4998C17.0248 1.04166 18.9582 2.97499 18.9582 7.49999V12.5C18.9582 17.025 17.0248 18.9583 12.4998 18.9583ZM7.49984 2.29166C3.65817 2.29166 2.2915 3.65832 2.2915 7.49999V12.5C2.2915 16.3417 3.65817 17.7083 7.49984 17.7083H12.4998C16.3415 17.7083 17.7082 16.3417 17.7082 12.5V7.49999C17.7082 3.65832 16.3415 2.29166 12.4998 2.29166H7.49984Z" fill="#808080"/>
-<path d="M13.125 8.125H6.875C6.53333 8.125 6.25 7.84167 6.25 7.5C6.25 7.15833 6.53333 6.875 6.875 6.875H13.125C13.4667 6.875 13.75 7.15833 13.75 7.5C13.75 7.84167 13.4667 8.125 13.125 8.125Z" fill="#808080"/>
-<path d="M13.125 13.125H6.875C6.53333 13.125 6.25 12.8417 6.25 12.5C6.25 12.1583 6.53333 11.875 6.875 11.875H13.125C13.4667 11.875 13.75 12.1583 13.75 12.5C13.75 12.8417 13.4667 13.125 13.125 13.125Z" fill="#808080"/>
-</svg>
-`,
-optionIcon:`<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M12.4998 18.9583H7.49984C2.97484 18.9583 1.0415 17.025 1.0415 12.5V7.49999C1.0415 2.97499 2.97484 1.04166 7.49984 1.04166H12.4998C17.0248 1.04166 18.9582 2.97499 18.9582 7.49999V12.5C18.9582 17.025 17.0248 18.9583 12.4998 18.9583ZM7.49984 2.29166C3.65817 2.29166 2.2915 3.65832 2.2915 7.49999V12.5C2.2915 16.3417 3.65817 17.7083 7.49984 17.7083H12.4998C16.3415 17.7083 17.7082 16.3417 17.7082 12.5V7.49999C17.7082 3.65832 16.3415 2.29166 12.4998 2.29166H7.49984Z" fill="#808080"/>
-<path d="M13.125 8.125H6.875C6.53333 8.125 6.25 7.84167 6.25 7.5C6.25 7.15833 6.53333 6.875 6.875 6.875H13.125C13.4667 6.875 13.75 7.15833 13.75 7.5C13.75 7.84167 13.4667 8.125 13.125 8.125Z" fill="#808080"/>
-<path d="M13.125 13.125H6.875C6.53333 13.125 6.25 12.8417 6.25 12.5C6.25 12.1583 6.53333 11.875 6.875 11.875H13.125C13.4667 11.875 13.75 12.1583 13.75 12.5C13.75 12.8417 13.4667 13.125 13.125 13.125Z" fill="#808080"/>
-</svg>
-`,
- adminIcon:`<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M12.4998 18.9583H7.49984C2.97484 18.9583 1.0415 17.025 1.0415 12.5V7.49999C1.0415 2.97499 2.97484 1.04166 7.49984 1.04166H12.4998C17.0248 1.04166 18.9582 2.97499 18.9582 7.49999V12.5C18.9582 17.025 17.0248 18.9583 12.4998 18.9583ZM7.49984 2.29166C3.65817 2.29166 2.2915 3.65832 2.2915 7.49999V12.5C2.2915 16.3417 3.65817 17.7083 7.49984 17.7083H12.4998C16.3415 17.7083 17.7082 16.3417 17.7082 12.5V7.49999C17.7082 3.65832 16.3415 2.29166 12.4998 2.29166H7.49984Z" fill="#808080"/>
-<path d="M13.125 8.125H6.875C6.53333 8.125 6.25 7.84167 6.25 7.5C6.25 7.15833 6.53333 6.875 6.875 6.875H13.125C13.4667 6.875 13.75 7.15833 13.75 7.5C13.75 7.84167 13.4667 8.125 13.125 8.125Z" fill="#808080"/>
-<path d="M13.125 13.125H6.875C6.53333 13.125 6.25 12.8417 6.25 12.5C6.25 12.1583 6.53333 11.875 6.875 11.875H13.125C13.4667 11.875 13.75 12.1583 13.75 12.5C13.75 12.8417 13.4667 13.125 13.125 13.125Z" fill="#808080"/>
-</svg>
-`,
-
-tagIcon:`<svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M12.4998 18.9583H7.49984C2.97484 18.9583 1.0415 17.025 1.0415 12.5V7.49999C1.0415 2.97499 2.97484 1.04166 7.49984 1.04166H12.4998C17.0248 1.04166 18.9582 2.97499 18.9582 7.49999V12.5C18.9582 17.025 17.0248 18.9583 12.4998 18.9583ZM7.49984 2.29166C3.65817 2.29166 2.2915 3.65832 2.2915 7.49999V12.5C2.2915 16.3417 3.65817 17.7083 7.49984 17.7083H12.4998C16.3415 17.7083 17.7082 16.3417 17.7082 12.5V7.49999C17.7082 3.65832 16.3415 2.29166 12.4998 2.29166H7.49984Z" fill="#808080"/>
-<path d="M13.125 8.125H6.875C6.53333 8.125 6.25 7.84167 6.25 7.5C6.25 7.15833 6.53333 6.875 6.875 6.875H13.125C13.4667 6.875 13.75 7.15833 13.75 7.5C13.75 7.84167 13.4667 8.125 13.125 8.125Z" fill="#808080"/>
-<path d="M13.125 13.125H6.875C6.53333 13.125 6.25 12.8417 6.25 12.5C6.25 12.1583 6.53333 11.875 6.875 11.875H13.125C13.4667 11.875 13.75 12.1583 13.75 12.5C13.75 12.8417 13.4667 13.125 13.125 13.125Z" fill="#808080"/>
-</svg>
-`,
-
-embedIcon: `<svg width="20" height="23" viewBox="0 0 20 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M1.6665 20.3688H18.3332" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 1.8517C11.3333 2.44425 12.8333 2.44425 14.1667 1.8517V4.62926C12.8333 5.2218 11.3333 5.2218 10 4.62926V1.8517Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 4.62927V7.40683" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M14.1668 7.40683H5.8335C4.16683 7.40683 3.3335 8.33268 3.3335 10.1844V20.3688H16.6668V10.1844C16.6668 8.33268 15.8335 7.40683 14.1668 7.40683Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M3.81689 11.1102H16.1836" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M6.6582 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M9.9917 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M13.3252 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-</svg>
-`,
-
-sendLimitIcon: `<svg width="20" height="23" viewBox="0 0 20 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M1.6665 20.3688H18.3332" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 1.8517C11.3333 2.44425 12.8333 2.44425 14.1667 1.8517V4.62926C12.8333 5.2218 11.3333 5.2218 10 4.62926V1.8517Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 4.62927V7.40683" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M14.1668 7.40683H5.8335C4.16683 7.40683 3.3335 8.33268 3.3335 10.1844V20.3688H16.6668V10.1844C16.6668 8.33268 15.8335 7.40683 14.1668 7.40683Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M3.81689 11.1102H16.1836" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M6.6582 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M9.9917 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M13.3252 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-</svg>
-`,
-rondwellPlusIcon: `<svg width="20" height="23" viewBox="0 0 20 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-<path d="M1.6665 20.3688H18.3332" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 1.8517C11.3333 2.44425 12.8333 2.44425 14.1667 1.8517V4.62926C12.8333 5.2218 11.3333 5.2218 10 4.62926V1.8517Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M10 4.62927V7.40683" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M14.1668 7.40683H5.8335C4.16683 7.40683 3.3335 8.33268 3.3335 10.1844V20.3688H16.6668V10.1844C16.6668 8.33268 15.8335 7.40683 14.1668 7.40683Z" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M3.81689 11.1102H16.1836" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linecap="round" stroke-linejoin="round"/>
-<path d="M6.6582 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M9.9917 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-<path d="M13.3252 11.1102V20.3688" stroke="#808080" stroke-width="1.5" stroke-miterlimit="10" stroke-linejoin="round"/>
-</svg>
-`
-
+	function checkScreenSize() {
+		if (browser) {
+			isMobile = window.innerWidth < 768;
+		}
 	}
 
+	onMount(() => {
+		checkScreenSize();
+		window.addEventListener('resize', checkScreenSize);
+		return () => window.removeEventListener('resize', checkScreenSize);
+	});
+
+	const themes = {
+		default: 'linear-gradient(180deg, #DBD4F1 0%, #DBE5F5 17%, #F4F5F6 35%)'
+	};
+
+	let selectedTheme = themes.default;
+
+	function getThemeForRoute(_: string | null): string {
+		return themes.default;
+	}
 
 	function updateCollectionSubMenu(collectionId: string) {
 		subMenuItems.set([
 			{
 				label: 'Events',
 				icon: collectionIcon.eventIcon,
-				nav: `/collection/${collectionId}`
+				nav: `/collection/${collectionId}/events`
 			},
 			{
 				label: 'People',
@@ -173,76 +116,21 @@ rondwellPlusIcon: `<svg width="20" height="23" viewBox="0 0 20 23" fill="none" x
 		]);
 	}
 
-	function updateSettingsSubMenu(collectionId: string) {
-	settingsSubMenuItems.set([
-		{ label: 'Display',
-		icon: settingIcon.displayIcon, nav: `/collection/${collectionId}/settings` },
-		{ label: 'Payment', 
-		icon: settingIcon.paymentIcon, nav: `/collection/${collectionId}/settings/payment` },
-		{ label: 'Options',
-		icon: settingIcon.optionIcon, nav: `/collection/${collectionId}/settings/options` },
-		{ label: 'Admins', 
-		icon: settingIcon.adminIcon, nav: `/collection/${collectionId}/settings/admins` },
-		{ label: 'Tags',
-		icon: settingIcon.tagIcon, 
-		 nav: `/collection/${collectionId}/settings/tags` },
-		{ label: 'Embed',
-		icon: settingIcon.embedIcon, nav: `/collection/${collectionId}/settings/embed` },
-		{ label: 'Send limit',
-		icon: settingIcon.sendLimitIcon, nav: `/collection/${collectionId}/settings/send-limit` },
-		{ label: 'Rondwell Plus',
-		icon: settingIcon.rondwellPlusIcon, nav: `/collection/${collectionId}/settings/rondwell-plus` }
+	$: path = $page.url.pathname;
+	$: collectionId = $page.params.id;
 
-	]);
-}
+	$: if (collectionId) {
+		selectedTheme = getThemeForRoute(path);
 
-
-	$: {
-	const path = $page.url.pathname;
-
-	const collectionMatch = path.match(/^\/collection\/([^/]+)/);
-
-	if (collectionMatch) {
-		const collectionId = collectionMatch[1];
-
-		// ---- Level 1 submenu (Events, People, Settings...)
 		updateCollectionSubMenu(collectionId);
+
 		const currentMenu = get(subMenuItems);
+
 		const active = currentMenu.find((item) => path.startsWith(item.nav))?.label || 'Events';
+
 		activeSubItem.set(active);
 		showSubMenu.set(true);
-
-		// ---- Level 2 submenu (ONLY when in settings)
-		if (path.startsWith(`/collection/${collectionId}/settings`)) {
-			updateSettingsSubMenu(collectionId);
-			showSettingsSubMenu.set(true);
-			// goto(`/collection/${collectionId}/settings/display`, { replaceState: true });
-
-			const settingsMenu = get(settingsSubMenuItems);
-			const activeSettings =
-				settingsMenu.find((item) => path.startsWith(item.nav))?.label || 'Display';
-
-			activeSettingsItem.set(activeSettings);
-
-
-		} else {
-			showSettingsSubMenu.set(false);
-			settingsSubMenuItems.set([]);
-			activeSettingsItem.set('');
-		}
-
-	} else {
-		// Outside collection routes
-		showSubMenu.set(false);
-		subMenuItems.set([]);
-		activeSubItem.set('');
-
-		showSettingsSubMenu.set(false);
-		settingsSubMenuItems.set([]);
-		activeSettingsItem.set('');
 	}
-}
-
 </script>
 
 <slot />
