--- src/routes/(app)/(main)/collection/[id]/+page.ts
@@ -1,5 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-
-export function load({ params }) {
-	throw redirect(302, `/collection/${params.id}/events`);
-}
