+++ src/routes/(app)/(main)/collection/[id]/components/SelectDropdown.svelte
@@ -54,7 +54,7 @@
     "
 	>
 		<span class={value ? 'text-gray-700 dark:text-gray-200' : 'text-gray-400'}>
-			{ value ?? placeholder}
+			{value ?? placeholder}
 		</span>
 
 		<!-- Chevron -->
