--- src/routes/(app)/(main)/collection/[id]/settings/+page.ts
@@ -1,6 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-import type { PageLoad } from './$types';
-
-export const load: PageLoad = ({ params }) => {
-	throw redirect(302, `/collection/${params.id}/settings/display`);
-};
