+++ src/routes/(app)/(main)/collection/[id]/settings/components/Admins.svelte
@@ -1,7 +1,5 @@
 <script lang="ts">
-	import SubscriberDetailModal from "../../people/SubscriberDetailModal.svelte";
-
-	
+	import SubscriberDetailModal from '../../people/SubscriberDetailModal.svelte';
 
 	// Mock data
 
@@ -84,42 +82,7 @@
 
 <section class="p-4">
 	<!-- Header -->
-	<div class="hidden sm:justify-between lg-block md:mb-10 md:flex">
-		<div class="flex items-center justify-between">
-			<img src="/tech-icon.svg" alt="icon" class="h-7 w-7" />
-			<h1 class="ml-2 lg:text-2xl text-md">Business Collection</h1>
-		</div>
-		<button
-			class="flex items-center gap-2 w-fit rounded-md bg-[#DCE4EE] px-3 whitespace-nowrap md:text-sm text-[#5D646F]"
-		>
-			<span class="md:text-sm">Calendar Page</span>
-			<svg
-				width="16"
-				height="16"
-				viewBox="0 0 16 16"
-				fill="none"
-				xmlns="http://www.w3.org/2000/svg"
-			>
-				<path
-					d="M1.24306 6.4387C1.40611 5.40243 2.12888 4.62786 3.0931 4.47826L9.69034 3.43935L9.8408 3.42097C10.5948 3.35739 11.3249 3.72187 11.7721 4.3912C12.2195 5.06073 12.3131 5.92902 12.0244 6.68931L11.9618 6.83923L9.01457 13.3413L9.01326 13.3411C8.84871 13.7088 8.61528 14.0066 8.33157 14.2308C7.8792 14.5883 7.31432 14.7405 6.72781 14.6481C5.77143 14.4963 5.05093 13.7247 4.89305 12.6922L4.591 10.7138C4.53659 10.3578 4.3245 10.0403 4.02548 9.86912L4.02494 9.8683L2.3872 8.94152C1.53287 8.45922 1.08026 7.47484 1.24306 6.4387ZM2.76439 5.88928C2.52769 6.07636 2.39626 6.36366 2.35324 6.63719L2.35378 6.638C2.28828 7.04462 2.40721 7.57568 2.91465 7.86476L4.55152 8.78851L4.66279 8.85692C5.17357 9.19045 5.53909 9.73754 5.67294 10.3689L5.69839 10.5051L6.00044 12.4835L6.00098 12.4843C6.09306 13.0997 6.52266 13.3845 6.9 13.4426C7.27991 13.5012 7.75745 13.3662 8.00792 12.8142L10.9559 6.31145C11.1348 5.91836 11.1012 5.47237 10.8636 5.11651C10.6258 4.7606 10.2457 4.58735 9.84651 4.65089L9.84575 4.6515L3.24253 5.68148C3.04601 5.71271 2.88934 5.79056 2.76439 5.88928Z"
-					fill="#5D646F"
-					stroke="#5D646F"
-					stroke-width="0.37461"
-				/>
-				<rect
-					x="7.25931"
-					y="8.68484"
-					width="3.5114"
-					height="1.15881"
-					rx="0.579404"
-					transform="rotate(144 7.25931 8.68484)"
-					fill="#5D646F"
-					stroke="#5D646F"
-					stroke-width="0.37461"
-				/>
-			</svg>
-		</button>
-	</div>
+
 	<div class="mb-4 flex items-center justify-between">
 		<h2 class="align-center flex justify-center gap-2 text-lg font-semibold">
 			<span> Admins </span>
@@ -136,7 +99,7 @@
 
 	<!-- Description -->
 	<p class="mb-4 text-sm text-[#737577]">
-		Admins have full access to the calendar and can approve events. 
+		Admins have full access to the calendar and can approve events.
 	</p>
 
 	<div class="mb-4 gap-3">
@@ -145,12 +108,12 @@
 				type="text"
 				bind:value={searchQuery}
 				placeholder="Search"
-				class="mb-2 h-[43px] w-full rounded-lg bg-[#FFFFFF] py-2 pr-4 pl-10 text-[#C5C6C6] focus:ring-0 focus:outline-none"
+				class="mb-2 h-[43px] w-full rounded-lg bg-[#FFFFFF] py-2 pl-10 pr-4 text-[#C5C6C6] focus:outline-none focus:ring-0"
 			/>
-			<span class="absolute top-2.5 left-3 text-gray-400">
+			<span class="absolute left-3 top-2.5 text-gray-400">
 				<img src="/search-favorite.png" alt="search icon" class="h-5 w-5" />
 			</span>
-			<span class="absolute top-2.5 right-0 border-l px-4 py-1.5 text-gray-400">
+			<span class="absolute right-0 top-2.5 border-l px-4 py-1.5 text-gray-400">
 				<img src="/search-download.svg" alt="" class="h-4 w-4" />
 			</span>
 		</div>
